+++ Model/ModelAPI.cs
@@ -20,7 +20,7 @@ namespace Model
             Circles.Clear();
             foreach(IBall ball in _logicAPI.GetAllBalls())
             {
-                ICircle c = ICircle.CreateCircle(ball.PosX, ball.PosY, ball.Radius);
+                ICircle c = ICircle.CreateCircle((int)ball.PosX, (int)ball.PosY, (int)ball.Radius);
                 Circles.Add(c);                                  //Ponizej dodajemy metode, ktora bedzie wywolywana za kazdym razem, gdy ball zglosi PropertyChanged
                 ball.PropertyChanged += c.UpdateCircle!;         //wykrzyknik nie jest konieczny, to tylko mowi kompilatorowi ze metoda UpdateCircle nie bedzie NULLem
             }
